requirements.txt
fastmcp
aiosqlite
python-dotenv
starlette
langgraph
langchain-mcp-adapters
langchain-gigachat
